First, hacky attempt at a parallel evaluator
This adds a "nix search" command equivalent to "nix-env -qaP
--drv-path <nixpkgs>" that evaluates attributes in parallel.

Evaluation is made thread-safe (modulo a gazillion remaining races) by
reappropriating the black hole mechanism: if forceValue() encounters a
value marked as a black hole (meaning another thread is busy
evaluating it), it spins until the value is no longer a black hole.

Results for N=1..8 running

  CORES=$N nix search

on a i7-6700HQ (4 cores, 8 hyperthreads):

  N=1: 11.1112
  N=2:  6.7378
  N=3:  5.5765
  N=4:  4.6949
  N=5:  4.5008
  N=6:  4.3608
  N=7:  4.1978
  N=8:  4.0962

Thus, as to be expected for such a naive approach, there is a lot of
contention.

Note: this doesn't work with Boehm GC currently.