Add an aws:// store type that supports building
aws:// is an extension of s3:// that implements
buildDerivation(). This allows it to be used as a remote builder, e.g.

  $ nix-build -j0 -E \
      --builders 'aws://nix-cache?region=eu-west-1&sqs-queue=nix-build-queue' \
      'with import <nixpkgs> {}; patchelf.overrideDerivation (attrs: { x = 3; })'
  building '/nix/store/2aw78n4p60aca6l09r5x0qmfijpy48fa-patchelf-0.9.drv' on 'aws://nix-cache'...
  copying 1 paths...
  copying path '/nix/store/fjqfyiarhaspklhpv37bkwr81x4wg38l-patchelf-0.9' from 'aws://nix-cache'...

  real    0m7.961s

This will use the S3 bucket 'nix-cache' to store a binary cache
holding build inputs and outputs, and the SQS queue 'nix-build-queue'
to communicate with the actual builder (which typically is an EC2 spot
instance automatically provisioned depending on the queue length, but
it can be any machine). It also creates a temporary SQS queue to await
the build status. (In the future this queue may also be used to
propagate log output.)

Derivations are processed by running:

  $ nix process-aws-build-queue 'aws://nix-cache?region=eu-west-1&sqs-queue=nix-build-queue'

This program fetches derivation messages from the SQS queue, copies
the derivation inputs from S3, executes the build locally, uploads the
derivation outputs to S3, then wakes up the client via the temporary
SQS queue.