Start using the Rust StorePath type directly in C++
This involves some hackery to allow Rust values to be used directly in
C++, via a helper type whose destructor calls the Rust type's drop()
function. The main issue is the dynamic nature of C++ move semantics:
when we move a Rust value, then we should not call the drop function
on the original value. So when we move a value, we set the original
value to bitwise zero, and the destructor only calls drop() if the
value is not bitwise zero. This should be sufficient for most types.