add support for bespoke ownership and mode
This patch adds support for "ownershipUser" and "ownershipGroup" Nix
expression parameters to govern the user and group values for package
contents.  Setting these values has the implicit effect of removing
access for the "other" bits, and setting "ownershipUser" in the absence
of "ownershipGroup" also has the effect of clearing the "group" bits.

For example, these would be the directory modes:

                 if setting ownershipUser alone: dr-x------
                if setting ownershipGroup alone: dr-xr-x---
 if setting both ownershipUser & ownershipGroup: dr-xr-x---